refactor: user 클래스 분리 및 구입 금액 테스트 코드 구현
